exchange Long -> int
Long타입이 프로젝트 내에서 변환하기 힘들거같다는 의견을 수렴하여 int로 변경

+++ src/main/java/com/bringup/company/user/entity/Salary.java
@@ -13,10 +13,10 @@ public class Salary {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private int id;
 
     @Column(name = "company_id", nullable = false)
-    private Long companyId;
+    private int companyId;
 
     @Column(name = "position", nullable = false, length = 10)
     private String position;
